oca_backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
supabase==2.3.1
python-dotenv==1.0.1
pydantic==2.6.3
pydantic-settings==2.2.1
gunicorn==21.2.0
httpx==0.27.0
